% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obs_day_maxTem}
\alias{obs_day_maxTem}
\title{Plot day maximum temperature observations over China.}
\usage{
obs_day_maxTem(
  date,
  threshold = -100,
  limit = c(18, 72, 54, 136),
  table = TRUE
)
}
\arguments{
\item{date, }{date string, like "20160801"}

\item{threshold, }{temperature larger or equal than threshold will be displayed}

\item{limit, }{region limit, (minLat,minLon,maxLat,maxLon)}

\item{table, }{display topest n record or not, default = TRUE.}
}
\value{
ggplot2 plot
}
\description{
Plot day maximum temperature observations over China.
}
\examples{
   obs_day_maxTem("20160811")

}
