% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obs_day_rain}
\alias{obs_day_rain}
\title{Plot day accumulated precipitation observations over China.}
\usage{
obs_day_rain(
  date,
  hour = 8,
  threshold = 0.1,
  limit = c(18, 72, 54, 136),
  table = TRUE
)
}
\arguments{
\item{date, }{date string, like "20160801"}

\item{hour, }{observation hour, 8 or 20.}

\item{threshold, }{24h precipitation larger or equal than threshold will be displayed}

\item{limit, }{region limit, (minLat,minLon,maxLat,maxLon)}

\item{table, }{display topest n record or not, default = TRUE.}
}
\value{
ggplot2 plot
}
\description{
Plot day accumulated precipitation observations over China.
}
\examples{
    obs_day_rain("20160814")

}
