% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obs_rain1h_series}
\alias{obs_rain1h_series}
\title{Plot station one hour rainfall time series.}
\usage{
obs_rain1h_series(staIds, startTime = NULL, endTime = NULL)
}
\arguments{
\item{staIds}{: station ids, like "54511,54416,54406"}

\item{startTime}{: start and end time string, like "2016071900"}

\item{endTime}{: end time string, like "2016072100"}
}
\value{
ggplot plot.
}
\description{
\preformatted{Plot station one hour rainfall observation time series.
}
}
\examples{
    obs_rain1h_series("54511,54416,54406") # display latest 72 hours rainfall.
    obs_rain1h_series("54511,54416,54406", startTime="2016071900",endTime="2016072100")

}
