% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{surf_1h_obs_pre_series}
\alias{surf_1h_obs_pre_series}
\title{Title}
\usage{
surf_1h_obs_pre_series(staIds, startTimeStr, nhours = 36)
}
\arguments{
\item{staIds}{}

\item{startTimeStr}{}

\item{nhours}{}
}
\value{

}
\description{
Title
}
\examples{
   g <- surf_1h_obs_pre_series("54399,54511", startTimeStr="2020020500", nhours=36)
   
}
